+++ config/messages.py
@@ -638,6 +638,7 @@ if ("Mám průměr pod 2.0")                                           \n  retur
     timeout_wars_user = "Uživatel {user} byl umlčen na {time:.0f} minut."
     timeout_wars_message_delete = "Uživatel {user} byl přistižen při mazání zpráv. Byl proto umlčen na {time:.0f} minut."
     timeout_wars_user_immunity = "Uživatel {user} má ještě imunitu na {time:.0f} sekund."
+    timeout_wars_reason = "timeoutwars"
 
     # MODERATION
     slowmode_time = "Délka prodlevy v sekundách (vyber v autocomplete)"
